Allow player to toggle avoided items. Mark them in sidebar.
Previously, if the player dropped something, they would avoid it in auto play.
Now, if the player targets an item, they can now toggle this avoidance on/off.
This is done with the 'v' key.
If an item is marked to be avoided, it's displayed in the sidebar with a red
background to the ':'.