Add interactive auto-explore which lets you cycle through possible targets.
The current auto-explore works great most of the time, but when it
doesn't, it can cost you dozens of turns, or even get you into serious
trouble.

Press 'w' to cycle through all items (that aren't being avoided), and a
set of spots at the boundary of unexplored space.
They're ordered from nearest to the player to farthest.
The unexplored spots are chosen with a very simplistic method. But it
works reasonably well. The important thing is they're just suggestions.
You can always just skip or adjust them.

Press comma (,) to shorten the path by a space. This is really handy since
the "unexplored" suggested targets frequently overshoot the best spot a bit.

If you're in "keys move cursor" mode, before pressing either of these keys,
you'll stay in it, so you can adjust the target with movement keys, too.